[lib] Restructured the library for expansion
- Add: Added `.image` subpackage.
- Change: Moved `.image` module to `.image.common`.
